refactor: split out prompt models
