+++ smarter/smarter/apps/prompt/models/__init__.py
@@ -1,0 +1,13 @@
+from .prompt import Prompt
+from .prompt_helper import PromptHelper
+from .prompt_history import PromptHistory
+from .prompt_plugin_usage import PromptPluginUsage
+from .prompt_tool_call import PromptToolCall
+
+__all__ = [
+    "Prompt",
+    "PromptHelper",
+    "PromptPluginUsage",
+    "PromptToolCall",
+    "PromptHistory",
+]
